rename MarkdownNotebook to NoteWorkspace
